Crate opentelemetry_proto

This crate contains generated files from opentelemetry-proto repository and transformation between types from generated files and types defined in opentelemetry

Based on the build tool needed, users can choose to generate files using tonic or grpcio.

Feature flags

Opentelemetry-proto includes a set of feature flags to avoid pull in unnecessary dependencies. The following is the full list of currently supported features:

Signals

trace: generate types that used in traces. Currently supports gen-tonic.

metrics: generate types that used in metrics. Currently supports gen-tonic.

logs: generate types that used in logs. Currently supports gen-tonic.

zpages: generate types that used in zPages. Currently only tracez related types will be generated. Currently supports gen-tonic.

Creates used to generate files

gen-tonic-messages: generate rs files using tonic and prost.

gen-tonic: adding tonic transport to “`gen-tonic-messages”

Misc

full: enabled all features above.

By default, no feature is enabled.

Modules

tonic

Generated files using tonic and prost

transform