opentelemetry_proto::tonic::metrics

Module v1

Modules

exemplar

Nested message and enum types in Exemplar.

exponential_histogram_data_point

Nested message and enum types in ExponentialHistogramDataPoint.

metric

Nested message and enum types in Metric.

number_data_point

Nested message and enum types in NumberDataPoint.

summary_data_point

Nested message and enum types in SummaryDataPoint.

Structs

Exemplar

A representation of an exemplar, which is a sample input measurement. Exemplars also hold information about the environment when the measurement was recorded, for example the span and trace ID of the active span when the exemplar was recorded.

ExponentialHistogram

ExponentialHistogram represents the type of a metric that is calculated by aggregating as a ExponentialHistogram of all reported double measurements over a time interval.

ExponentialHistogramDataPoint

ExponentialHistogramDataPoint is a single data point in a timeseries that describes the time-varying values of a ExponentialHistogram of double values. A ExponentialHistogram contains summary statistics for a population of values, it may optionally contain the distribution of those values across a set of buckets.

Gauge

Gauge represents the type of a scalar metric that always exports the “current value” for every data point. It should be used for an “unknown” aggregation.

Histogram

Histogram represents the type of a metric that is calculated by aggregating as a Histogram of all reported measurements over a time interval.

HistogramDataPoint

HistogramDataPoint is a single data point in a timeseries that describes the time-varying values of a Histogram. A Histogram contains summary statistics for a population of values, it may optionally contain the distribution of those values across a set of buckets.

Metric

Defines a Metric which has one or more timeseries. The following is a brief summary of the Metric data model. For more details, see:

MetricsData

MetricsData represents the metrics data that can be stored in a persistent storage, OR can be embedded by other protocols that transfer OTLP metrics data but do not implement the OTLP protocol.

NumberDataPoint

NumberDataPoint is a single data point in a timeseries that describes the time-varying scalar value of a metric.

ResourceMetrics

A collection of ScopeMetrics from a Resource.

ScopeMetrics

A collection of Metrics produced by an Scope.

Sum

Sum represents the type of a scalar metric that is calculated as a sum of all reported measurements over a time interval.

Summary

Summary metric data are used to convey quantile summaries, a Prometheus (see: https://prometheus.io/docs/concepts/metric_types/#summary) and OpenMetrics (see: https://github.com/prometheus/OpenMetrics/blob/4dbf6075567ab43296eed941037c12951faafb92/protos/prometheus.proto#L45) data type. These data points cannot always be merged in a meaningful way. While they can be useful in some applications, histogram data points are recommended for new applications. Summary metrics do not have an aggregation temporality field. This is because the count and sum fields of a SummaryDataPoint are assumed to be cumulative values.

SummaryDataPoint

SummaryDataPoint is a single data point in a timeseries that describes the time-varying values of a Summary metric. The count and sum fields represent cumulative values.

Enums

AggregationTemporality

AggregationTemporality defines how a metric aggregator reports aggregated values. It describes how those values relate to the time interval over which they are aggregated.

DataPointFlags

DataPointFlags is defined as a protobuf ‘uint32’ type and is to be used as a bit-field representing 32 distinct boolean flags. Each flag defined in this enum is a bit-mask. To test the presence of a single flag in the flags of a data point, for example, use an expression like: