opentelemetry_proto::tonic::metrics::v1::number_data_point

Enum Value

pub enum Value {
    AsDouble(f64),
    AsInt(i64),
}

The value itself. A point is considered invalid when one of the recognized value fields is not present inside this oneof.

Variants

AsDouble(f64)

AsInt(i64)

Implementations

impl Value

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Value>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.